fix: move useTimer to its folder
